+++ apps/kyb-app/src/domains/collection-flow/collection-flow.api.ts
@@ -139,6 +139,16 @@ export const syncContext = async (context: CollectionFlowContext) => {
   return result.json();
 };
 
+export const finalSubmissionRequest = async () => {
+  const result = await request.post('collection-flow/final-submission', {
+    json: {
+      eventName: 'COLLECTION_FLOW_FINISHED',
+    },
+  });
+
+  return result.json();
+};
+
 export const fetchDocumentsByIds = async (ids: string[]) => {
   const result = await request.get('collection-flow/files', {
     searchParams: {