+++ services/workflows-service/src/app.module.ts
@@ -50,6 +50,7 @@ import { NotionModule } from '@/notion/notion.module';
 import { SecretsManagerModule } from '@/secrets-manager/secrets-manager.module';
 import { NoteModule } from '@/note/note.module';
 import { DocumentModule } from './document/document.module';
+import { MerchantMonitoringModule } from './merchant-monitoring/merchant-monitoring.module';
 
 export const validate = async (config: Record<string, unknown>) => {
   const zodEnvSchema = z
@@ -91,6 +92,7 @@ export const validate = async (config: Record<string, unknown>) => {
     }),
     EventEmitterModule.forRoot(),
     UserModule,
+    MerchantMonitoringModule,
     WorkflowModule,
     WebhooksModule,
     NoteModule,