+++ services/workflows-service/src/collection-flow/controllers/collection-flow.no-user.controller.intg.test.ts
@@ -8,7 +8,7 @@ import request from 'supertest';
 import { WorkflowTokenRepository } from '@/auth/workflow-token/workflow-token.repository';
 import { WorkflowTokenService } from '@/auth/workflow-token/workflow-token.service';
 import { BusinessReportService } from '@/business-report/business-report.service';
-import { MerchantMonitoringClient } from '@/business-report/merchant-monitoring-client';
+import { MerchantMonitoringClient } from '@/merchant-monitoring/merchant-monitoring.client';
 import { BusinessRepository } from '@/business/business.repository';
 import { BusinessService } from '@/business/business.service';
 import { CollectionFlowService } from '@/collection-flow/collection-flow.service';