+++ services/workflows-service/src/workflow-defintion/workflow-definition.service.ts
@@ -250,6 +250,22 @@ export class WorkflowDefinitionService {
     return await this.workflowDefinitionRepository.updateById(id, args, projectIds, true);
   }
 
+  async createDemoWorkflowDefinition({
+    customerId,
+    userId,
+    workflowOverrides,
+  }: {
+    customerId: string;
+    userId?: string;
+    workflowOverrides?: Array<{ webPresenceReportId?: string }>;
+  }) {
+    return await this.workflowDefinitionRepository.createDemoWorkflowDefinition(
+      customerId,
+      userId,
+      workflowOverrides,
+    );
+  }
+
   async getByWorkflowRuntimeDataId(workflowRuntimeDataId: string, projectIds: TProjectIds) {
     return await this.workflowDefinitionRepository.findByWorkflowRuntimeDataId(
       workflowRuntimeDataId,