+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @ballerine/headless-example
 
+## 0.3.98
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.78
+  - @ballerine/workflow-browser-sdk@0.6.99
+
+## 0.3.97
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.98
+
+## 0.3.96
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.77
+  - @ballerine/workflow-browser-sdk@0.6.97
+
+## 0.3.95
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.76
+  - @ballerine/workflow-browser-sdk@0.6.96
+
 ## 0.3.94
 
 ### Patch Changes